+++ routes/users.js
@@ -29,76 +29,7 @@ const saltRounds = 16;
 const router = Router();
 const geocoder = NodeGeocoder(geocoderConfig);
 
-router
-  .route("/leaderboard/local")
-  .get(async (req, res) => {
-    const hasUserId = req.session.user && req.session.user._id;
 
-    const hasUserCountryCode =
-      req.session.user &&
-      req.session.user.geocode &&
-      req.session.user.geocode.countryCode;
-    const hasUserCity =
-      req.session.user &&
-      req.session.user.geocode &&
-      req.session.user.geocode.city;
-
-    let leaderboard;
-    try {
-      if (hasUserId)
-        leaderboard = await topNthLocalUsersByHighScore(
-          100,
-          req.session.user.geocode.countryCode,
-          req.session.user.geocode.city
-        );
-      else
-        leaderboard = await topNthLocalUsersByHighScore(100, "US", "New York");
-    } catch (error) {
-      return res.status(500).send("Internal Server Error:", error);
-    }
-
-    return res.render("leaderboard", {
-      title: "Local Leaderboard",
-      type: "local",
-      leaderboard,
-    });
-  })
-  .post(async (req, res) => {
-    // reserved for AJAX
-    var countryInput = req.body.countryInput;
-    var citySearchTerm = req.body.citySearchBar;
-
-    try {
-      const leaderboard = await topNthLocalUsersByHighScore(
-        100,
-        countryInput,
-        citySearchTerm
-      );
-      return res.send(leaderboard)
-    }
-    catch(error){
-      return res.render('leaderboard',{title: "Local Leaderboard", error:error})
-    }
-
-
-  });
-
-router.route("/leaderboard/global").get(async (req, res) => {
-  const hasUserId = req.session.user && req.session.user._id;
-
-  let leaderboard;
-  try {
-    leaderboard = await topNthGlobalUsersByHighScore(100);
-  } catch (error) {
-    //return res.status(500).send("Internal Server Error:", error);
-    return res.status(500).render(leaderboard,{title: "Global Leaderboard", error:error})
-  }
-
-  return res.render("leaderboard", {
-    title: "Global Leaderboard",
-    leaderboard,
-  });
-});
 
 router
   .route("/user")
@@ -192,104 +123,6 @@ router
     // req.session.user = { _id: newUser._id, username: newUser.username };
   
     //return res.redirect('/login');
-  });
-
-router.
-  route('/gamestart')
-  .get(async (req,res)=>{
-      res.render('game_start',{title: 'Quiz'});
-      })
-  .post(async(req,res)=>{
-      if(req.session.user){
-          req.session.questions = await getQuestions4User({
-            numberOfOptions : 4,
-            numberOfQuestions : 50,
-            countryCode : req.session.user['geocode.countryCode'],
-            city : req.session.user['geocode.city'],
-          });
-      }else{
-          req.session.questions =await getQuestions4Guest({
-            numberOfOptions : 4,
-            numberOfQuestions : 50,
-          } );
-      }
-      req.session.index = 0;
-      req.session.score = 0;
-      req.session.timer = 60
-      res.redirect('/users/gameplay');
-  });
-  
-router.
-  route('/gameplay')
-  .get(async (req, res) => {
-      let questions = req.session.questions;
-      let index = req.session.index;
-      let time = req.session.timer;
-      if(!req.session.score){
-          req.session.score=0;
-      }
-      if(questions.length<=index){
-          if(req.session.timer>0){
-              req.session.score = (req.session.score)*(60/(60-req.session.timer));
-          }
-          res.redirect('/users/gameresult');
-      }
-      let score = req.session.score;
-      return res.render('game_question', {title: 'Quiz',question : questions[index],index : index+1,score : score,time});
-      })
-  .post(async(req,res)=>{
-      let questions = req.session.questions;
-      let index = req.session.index;
-      const {options,timer} = req.body;
-      if(timer<=0){
-          if(options){
-              let correct = questions[index]['answer'];
-              if(questions[index]['options'][options] == correct){
-                  req.session.score = req.session.score + questions[index]['difficulty'];
-              }
-              req.session.timer = timer;
-          }
-          res.redirect('/users/gameresult');
-      }else{
-          let correct = questions[index]['answer'];
-          if(questions[index]['options'][options] == correct){
-              req.session.score = req.session.score + questions[index]['difficulty'];
-          }
-          req.session.index = index + 1;
-          req.session.timer = timer;
-          res.redirect('/users/gameplay')
-      }
-  });
-
-router.
-  route('/gameresult')
-  .get(async (req, res) => {
-      if(req.session.user){
-          let questions = req.session.questions;
-          let score = req.session.score;
-          let user =req.session.user;
-          for(let i=0;i<index;i++){
-              if(i<questions.length){
-                  let birdid = questions[i]['birdid'];
-                  await updatePlayerInfoById(user['_id'],{
-                    $pushLastQuestions: { birdid},
-                  });
-              }
-          }
-          delete req.session['questions'];
-          delete req.session['index'];
-          delete req.session['score'];
-          delete req.session['timer'];
-          res.render('game_end',{score});
-      }
-      else{
-          let score = req.session.score;
-          delete req.session['questions'];
-          delete req.session['index'];
-          delete req.session['score'];
-          delete req.session['timer'];
-          res.render('game_end',{score});
-      }
   });
 
 router
