+++ views/game_end.handlebars
@@ -1,5 +1,5 @@
 <h1>Results</h1>
 
 <p>Your score was {{score}}</p>
-<a id="LeaderBoardBtn" name="anchorBtn" href="/users/leaderboard/global">LeaderBoard</button>
+<a id="LeaderBoardBtn" name="anchorBtn" href="/leaderboard/global">LeaderBoard</button>
 <a id="HomepageBtn" name="anchorBtn" href="/">Homepage</button>
