+++ views/game_question.handlebars
@@ -2,7 +2,7 @@
             <h3>Score : <span id="score">{{score}}</span></h1>
             <h3> Question  <span id="questionNumber">{{index}}</span></h1>
 </div>
-    <form  id="questionform" name="questionform" action="/users/gameplay" method="POST">
+    <form  id="questionform" name="questionform" action="/game/gameplay" method="POST">
     <div id="tinmerContainer">
         <label for="timer" id="timerLabel">Time Remaining:</label><input type="text"  id="timer" name="timer" value={{time}} >
     </div>
