+++ views/game_start.handlebars
@@ -1,6 +1,6 @@
 <a href="/" id="homepageBtn" name="anchorBtn">Homepage</a>
 <div id="formContainer">
-    <form  id="gamestartForm" action="/users/gamestart" method="POST">
+    <form  id="gamestartForm" action="/game/gamestart" method="POST">
     <div id="gamestartContainer">
         <button id="startBtn" type="submit">Start Quiz</button>
     </div>
