+++ views/homepage.handlebars
@@ -20,13 +20,13 @@
     </div>
     <div class="container" id = "main_buttons">
         <div class = "play_button">
-            <a href="/users/gamestart" class="btn btn-link" id = "homepage_play_button">
+            <a href="/game/gamestart" class="btn btn-link" id = "homepage_play_button">
                 <img src="/static/images/play_button.png" alt="Play Button" class="play_button_image">
             </a>
         </div>
         
         <div class = "leaderboard_button">
-            <a href="/users/leaderboard/global" class="btn btn-link" id = "homepage_leaderboard_button">
+            <a href="/leaderboard/global" class="btn btn-link" id = "homepage_leaderboard_button">
             <img src="/static/images/leaderboard_button.png" alt="Leaderboard Button" class="leaderboard_button_image">
         </a>
     </div>
