+++ views/leaderboard.handlebars
@@ -1,7 +1,7 @@
 <div class="container">
   <h1 class="leaderboard-heading">{{title}}</h1>
-  <a class="btn btn-dark btn-lg" href="/users/leaderboard/local" role="button" id ="local_leaderboard_button">Local</a>
-  <a class="btn btn-dark btn-lg" href="/users/leaderboard/global" role="button" id ="global_leaderboard_button">Global</a>
+  <a class="btn btn-dark btn-lg" href="/leaderboard/local" role="button" id ="local_leaderboard_button">Local</a>
+  <a class="btn btn-dark btn-lg" href="/leaderboard/global" role="button" id ="global_leaderboard_button">Global</a>
   <br>
   <br>
   {{#if type}}
